+++ app/client/front/js/ov/HomeController.js
@@ -10,9 +10,19 @@
 
   /**
    * Defines the home page controller.
+   *
+   * @param {Object} promotedVideos Request results for promoted videos
+   * @param {Object} promotedVideos.data Request data
+   * @param {Array} promotedVideos.data.entities The list of promoted videos
    */
-  function HomeController($scope, $location, result) {
-    $scope.videos = result && result.data ? result.data.entities : [];
+  function HomeController($scope, $location, promotedVideos) {
+    if (promotedVideos && promotedVideos.data) {
+      $scope.videos = promotedVideos.data.filter(function(promotedVideo) {
+        return (promotedVideo);
+      });
+    } else
+      $scope.videos = [];
+
     $scope.search = {};
 
     $scope.searchSubmit = function() {
@@ -21,6 +31,6 @@
   }
 
   app.controller('HomeController', HomeController);
-  HomeController.$inject = ['$scope', '$location', 'result'];
+  HomeController.$inject = ['$scope', '$location', 'promotedVideos'];
 
 })(angular.module('ov.portal'));